server/requirements.txt
qrcode
Pillow
pynput
pyxinput
PyUserInput